chore: simplify recent changes from /simplify review
- App.tsx: drop unnecessary `typeof window` guard (Tauri always has
  window) and use lazy useState initializers so parseHash doesn't run
  on every render.
- gemini.ts: own the Gemini visibility helpers (`isGeminiModelAutoHidden`,
  `isGeminiModelVisible`) — they're Gemini domain logic, not generic
  preferences. Add `GEMINI_FREE_TIER` constant to replace stringly-typed
  tier checks. Drop the change-narrating block comment.
- Settings.tsx: parallelize the mount-time `exists()` + `fetchGeminiModels()`
  (they're independent), surface `expired`/`error` from the mount path
  so users get feedback instead of a silently empty model list, drop the
  outdated "Optimistic update" what-comment, and consume the new
  GEMINI_FREE_TIER constant.
