+++ src/lib/preferences.ts
@@ -35,24 +35,3 @@ export async function setGeminiModelVisible(modelId: string, visible: boolean): 
   await savePreferences(next);
   return next;
 }
-
-/**
- * Models that are not available on the user's current tier should default to
- * hidden so the dashboard and tray icon aren't pinned at 100% by buckets the
- * user can't actually use. Right now this only covers the known case from
- * issue #26 — free-tier users don't get Gemini Pro models.
- */
-export function isGeminiModelAutoHidden(modelId: string, tier: string): boolean {
-  if (tier === "free-tier" && modelId.toLowerCase().includes("pro")) return true;
-  return false;
-}
-
-export function isGeminiModelVisible(
-  modelId: string,
-  tier: string,
-  visibility: Record<string, boolean> | undefined
-): boolean {
-  const explicit = visibility?.[modelId];
-  if (typeof explicit === "boolean") return explicit;
-  return !isGeminiModelAutoHidden(modelId, tier);
-}
